+++ entity.h
@@ -7,27 +7,27 @@
 #include <SFML/Graphics/Color.hpp>
 #include "Vector.h"
 
-const float G = 6.67300e-11f;
+const double G = 6.67300e-11f;
 
 class Entity {
     private:
         std::string name;
         sf::Color color;
         Point3D position;
-        float mass;
-        float radius;
+        double mass;
+        double radius;
         Vector velocity;
         Vector acceleration;
     public:
-        Entity(std::string, std::array<uint8_t,3>, Point3D, float, float, Vector, Vector);
+        Entity(std::string, std::array<uint8_t,3>, Point3D, double, double, Vector, Vector);
         Point3D getPos();
         std::string getPlanetName();
-        float getMass();
-        float getRadius();
+        double getMass();
+        double getRadius();
         sf::Color getColor();
         void move();
-        void setVelocity(float, float, float);
-        void setAcceleration(float, float, float);
-        std::array<float, 3> attract(Entity&);
+        void setVelocity(double, double, double);
+        void setAcceleration(double, double, double);
+        std::array<double, 3> attract(Entity&);
         void updateMovement(std::vector<Entity*>&);
 };
